WIP: implement fading between colors on chord / color changes
So far I only implemented the color changing on the two biggest branches (both of which play the root note color). What I've found so far is discouraging, so I probably won't finish implementing this on the rest of the branches. Findings:

1) The fade jumps discretely -- it's not a smooth fade. This is probably because the low brightness LED colors don't have enough "color space" to fade smoothly from one color to another. Given the jumps, I feel like the fade effect is actually worse than no fade effect at all. Just rip the bandaid off. Fading between two similar colors is decent, but when the colors are very different, you notice the jumps.

Example of fade jumpiness:
https://www.youtube.com/watch?v=cVROhCQZGy8
https://www.youtube.com/watch?v=-eAXvy4TK2o

2) We now have occasional audio glitches. Perhaps the blending code is taxing. I could probably optimize my code a bit. Maybe this would solve the audio glitches. But maybe not?

Given the above, in particular (1), I don't think we should continue implementing fading.